feat(document): refactored code into diff comps
