+++ src/components/save-button.tsx
@@ -1,23 +1,18 @@
 import React from 'react';
 
 interface ISaveButtonProps {
-  docId: string | undefined;
+  docId?: string | undefined;
   saveChanges: Function;
 }
 
-const SaveButton: React.FC<ISaveButtonProps> = ({ docId, saveChanges }) => {
+const SaveButton: React.FC<ISaveButtonProps> = ({
+  docId = '',
+  saveChanges,
+}) => {
   return (
     <div className="save-container display-inline-block">
       <button
-        onClick={
-          docId
-            ? () => saveChanges()
-            : () => {
-                document
-                  .getElementById('modal-js-example')
-                  ?.classList.add('is-active');
-              }
-        }
+        onClick={() => saveChanges()}
         className={`saveBtn button is-small ${
           docId ? 'move-up' : 'save-margin'
         } is-primary ${!docId && 'js-modal-trigger'}`}
